logging: move Kube deprecation warning logger to cmd/prometheus
DedupDeprecationWarningLogger is the only thing in util/logging that needs
k8s.io/client-go, and cmd/prometheus is its only caller. Keeping it here made
every importer of util/logging link client-go: promql, scrape and
storage/remote all pulled in 79 Kubernetes packages they never call.

Move it next to its caller, where client-go is already imported. No behaviour
change; the deduplication tests move with it.
