+++ util/logging/dedupe.go
@@ -18,9 +18,6 @@ import (
 	"log/slog"
 	"sync"
 	"time"
-
-	"github.com/grafana/regexp"
-	"k8s.io/client-go/rest"
 )
 
 const (
@@ -136,45 +133,4 @@ func (d *Deduper) run() {
 			return
 		}
 	}
-}
-
-// deprecationRegex matches the format of Kubernetes API deprecation warnings:
-// See https://github.com/kubernetes/kubernetes/blob/da663405beb487d66c27a0220ea4073305ae9077/staging/src/k8s.io/apiserver/pkg/endpoints/deprecation/deprecation.go#L117.
-var deprecationRegex = regexp.MustCompile(`\S+ \S+ is deprecated in v\d+\.\d+\+`)
-
-// Even though deprecation warnings should be bounded in number, this safeguard should help prevent leaks.
-const maxDeprecationWarnings = 32
-
-// DedupDeprecationWarningLogger deduplicates Kube API deprecation warnings by message before logging them.
-// Inspired by https://github.com/kubernetes/kubernetes/blob/3edae6c1c49958fd10a708d9cc8c4c9e7f5fb6e8/staging/src/k8s.io/client-go/rest/warnings.go#L113
-type DedupDeprecationWarningLogger struct {
-	logger rest.WarningHandlerWithContext
-	lock   sync.Mutex
-	logged map[string]struct{}
-}
-
-func NewDedupDeprecationWarningLogger() *DedupDeprecationWarningLogger {
-	return &DedupDeprecationWarningLogger{
-		logger: rest.WarningLogger{},
-		logged: make(map[string]struct{}),
-	}
-}
-
-func (w *DedupDeprecationWarningLogger) HandleWarningHeaderWithContext(ctx context.Context, code int, agent, message string) {
-	if code != 299 || message == "" {
-		return
-	}
-
-	w.lock.Lock()
-	defer w.lock.Unlock()
-
-	if _, seen := w.logged[message]; seen {
-		return
-	}
-
-	if deprecationRegex.MatchString(message) && len(w.logged) < maxDeprecationWarnings {
-		w.logged[message] = struct{}{}
-	}
-
-	w.logger.HandleWarningHeaderWithContext(ctx, code, agent, message)
 }
