+++ util/logging/dedupe_test.go
@@ -15,8 +15,6 @@ package logging
 
 import (
 	"bytes"
-	"context"
-	"fmt"
 	"log/slog"
 	"strings"
 	"testing"
@@ -81,34 +79,4 @@ func TestDedupeConcurrent(t *testing.T) {
 	}
 
 	require.NotPanics(t, func() { concurrentWriteFunc() })
-}
-
-type fakeWarningLogger struct {
-	logs []string
-}
-
-func (fl *fakeWarningLogger) HandleWarningHeaderWithContext(_ context.Context, _ int, _, message string) {
-	fl.logs = append(fl.logs, message)
-}
-
-func TestDedupeDeprecationWarningLogger(t *testing.T) {
-	wl := DedupDeprecationWarningLogger{
-		logger: &fakeWarningLogger{},
-		logged: make(map[string]struct{}),
-	}
-
-	deprecationMessage := "v1 Endpoints is deprecated in v1.33+; use [discovery.k8s.io/v1](http://discovery.k8s.io/v1) EndpointSlice"
-	for range 10 {
-		wl.HandleWarningHeaderWithContext(context.Background(), 299, "", deprecationMessage)
-	}
-	require.Len(t, wl.logger.(*fakeWarningLogger).logs, 1)
-	require.Len(t, wl.logged, 1)
-	require.Equal(t, wl.logger.(*fakeWarningLogger).logs[0], deprecationMessage)
-
-	for i := range 10 {
-		wl.HandleWarningHeaderWithContext(context.Background(), 299, "", fmt.Sprintf("some other warning %d", i+1))
-	}
-	require.Len(t, wl.logger.(*fakeWarningLogger).logs, 11)
-	require.Len(t, wl.logged, 1)
-	require.Equal(t, "some other warning 10", wl.logger.(*fakeWarningLogger).logs[10])
 }
